eventflow-distribuido/eventflow-distribuido/ms-auth/requirements.txt
# Core Framework and Web Server
fastapi==0.103.2
uvicorn[standard]==0.23.2

# Database Driver
psycopg2-binary==2.9.9

# Authentication and Security (Versiones estables para evitar el error 500)
passlib==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# File Upload Support (Sin pin de versión para evitar el error de distribución)
python-multipart